Refactored according to feedback
