--- PL2/PL2-parser/src/test/java/nl/tudelft/pl2016gr2/parser/controller/FullGfaReaderTest.java
@@ -1,25 +1,0 @@
-package nl.tudelft.pl2016gr2.parser.controller;
-
-import static org.junit.Assert.assertEquals;
-
-import nl.tudelft.pl2016gr2.model.OriginalGraph;
-import org.junit.Test;
-
-/**
- * This class tests the {@link FullGfaReader} class.
- *
- * @author Cas
- */
-public class FullGfaReaderTest {
-
-  private static final String filename = "SMALL.gfa";
-
-  @Test
-  public void test() {
-    FullGfaReader fgr = new FullGfaReader(filename);
-    OriginalGraph og = fgr.getGraph();
-    assertEquals(og.getGenoms().size(), 11);
-    assertEquals(og.getNode(3).getGenomes().size(), 1);
-  }
-
-}
